requirements.txt
psutil==5.9.6
llama-cpp-python==0.2.11
streamlit==1.28.0
pandas==2.1.1
plotly==5.17.0
schedule==1.2.0
python-dotenv
transformers 
torch